Move `spend_sim.py` into tests (#18743)
* move spend_sim into tests

* use run_block_generator2() instead of get_name_puzzle_conditions() in CostLogger. It's only used to compute the cost of a spend bundle, in tests.
