+++ chia/_tests/wallet/cat_wallet/test_offer_lifecycle.py
@@ -5,7 +5,7 @@ from typing import Any, Optional
 import pytest
 from chia_rs import G2Element
 
-from chia.clvm.spend_sim import CostLogger, SimClient, SpendSim, sim_and_client
+from chia._tests.util.spend_sim import CostLogger, SimClient, SpendSim, sim_and_client
 from chia.types.blockchain_format.coin import Coin
 from chia.types.blockchain_format.program import Program
 from chia.types.blockchain_format.serialized_program import SerializedProgram
